backend/requirements.txt
Flask==2.3.3
python-socketio==5.9.0
python-engineio==4.7.1
eventlet==0.33.3